refactor sts region resolution
